requirements.txt
astroid==3.3.8
autopep8==2.3.2
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.8.2
branca==0.8.1
bs4==0.0.2
cached-property==2.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
coverage==7.6.12
cycler==0.12.1
dill==0.3.9
distlib==0.3.9
et_xmlfile==2.0.0
filelock==3.16.1
flake8==7.1.2
Flask==3.0.3
folium==0.19.5
fonttools==4.56.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.1.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==9.2.0
mysqlclient==2.2.4
nominatim==0.1
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
pip==24.2
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pylint==3.3.4
PyMySQL==1.1.1
pyparsing==3.2.1
PyPDF2==3.0.1
pytesseract==0.3.13
pytest==8.3.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
requests==2.32.3
setuptools==75.8.2
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.38
sqlalchemy-orm==1.2.10
tomlkit==0.13.2
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.27.1
virtualenvwrapper-win==1.2.7
Werkzeug==3.0.6
XlsxWriter==3.2.2
xyzservices==2025.1.0
yapf==0.43.0
speedtest==0.0.1
speedtest-cli==2.1.3
